cram_system/requirements.txt
apipkg==1.4
appdirs==1.4.3
appnope==0.1.0
beautifulsoup4==4.5.3
coreapi==2.3.0
coreschema==0.0.4
coverage==4.3.4
ddt==1.1.1
decorator==4.0.11
defusedxml==0.5.0
Django==1.11
django-allauth==0.31.0
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-filter==1.0.2
django-guardian==1.4.8
djangorestframework==3.6.2
drfdocs==0.0.11
execnet==1.4.1
gunicorn==19.7.1
httpie==0.9.9
ipython==6.0.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.10.2
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.10
oauthlib==2.0.2
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
py==1.4.33
Pygments==2.2.0
PyMySQL==0.7.11
pyparsing==2.2.0
pytest==3.0.7
pytest-cov==2.4.0
pytest-django==3.1.2
pytest-xdist==1.15.0
python3-openid==3.1.0
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
traitlets==4.3.2
uritemplate==3.0.0
wcwidth==0.1.7
